Revert "Slow Enemy movement"
This reverts commit 5d9e3ffeac6d199a07b210df051b3b58f7651b68.
