--- Assets/EnemyMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 79d61ec06a9a5334aade93448daf277f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
