+++ Assets/Prefabs/Turret.prefab
@@ -143,7 +143,6 @@ MonoBehaviour:
   target: {fileID: 0}
   range: 2
   enemyTag: Enemy
-  fireRate: 4
-  damageOverTime: 0
+  fireRate: 1
   bulletPrefab: {fileID: 6550542848919797217, guid: 92f82e83e321a36418da6d79d2926b8d, type: 3}
   firePoint: {fileID: 8140266263907771926}
